FLAC: support for "wasted bits".

+++ core/memory.h
@@ -3,7 +3,7 @@
 #include "core.h"
 
 /// Aligns \a offset to \a width (only for power of two \a width)
-inline uint align(uint width, uint offset) { assert((width&(width-1))==0); return (offset + (width-1)) & ~(width-1); }
+inline size_t align(size_t width, size_t offset) { assert((width&(width-1))==0); return (offset + (width-1)) & ~(width-1); }
 
 // C runtime memory allocation
 extern "C" void* malloc(size_t size) noexcept;
